--- src/main/java/com/example/demo/domain/Post.java
@@ -1,33 +1,0 @@
-package com.example.demo.domain;
-
-import com.example.demo.domain.global.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Post extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_id")
-    private Long id;
-
-    private String title;
-
-    private String content;
-
-    private String writer;
-
-    @Enumerated(EnumType.STRING)
-    private Type type;
-
-    private String password;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-}
