+++ src/main/java/com/example/demo/domain/User.java
@@ -1,6 +1,9 @@
 package com.example.demo.domain;
 
+import com.example.demo.domain.address.Address;
 import com.example.demo.domain.global.BaseTimeEntity;
+import com.example.demo.domain.post.Post;
+import com.example.demo.domain.volunteer.Volunteer;
 import jakarta.persistence.*;
 import lombok.*;
