+++ src/main/java/com/example/demo/domain/post/Type.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain;
+package com.example.demo.domain.post;
 
 public enum Type {
     HOBBY, MEMORY
